requirements.txt
openai
pandas
numpy
faker
python-dotenv